backend/requirements.txt
Flask==2.3.3
flask-cors==3.0.10
SQLAlchemy==1.4.50
alembic==1.10.4
pillow==10.1.0
pytesseract==0.3.11
pdf2image==1.16.3
PyPDF2==3.0.1
reportlab==4.1.0
werkzeug==2.3.7
python-multipart==0.0.6
typing_extensions
